biovision-ai/backend/requirements.txt
fastapi
uvicorn
tensorflow
numpy
pillow
motor
python-multipart
pydantic
python-dotenv
scikit-learn
matplotlib
